requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
WTForms==3.0.1
anyio==3.7.1
certifi==2023.5.7
dnspython==2.4.0
email-validator==2.0.0.post2
flask_mongoengine==1.0.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
sniffio==1.3.0
python-dotenv==1.0.0
flask_cors==3.0.10
flask_pymongo==2.3.0
PyJWT==2.8.0
passlib==1.7.4
